+++ packages/contracts/src/lib/Constants.sol
@@ -4,4 +4,4 @@ pragma solidity ^0.8.24;
 uint256 constant N_INPUTS = 5;
 uint256 constant N_OUTPUTS = 5;
 uint256 constant N_PUB = 3 + N_INPUTS + 3 * N_OUTPUTS + 1;
-uint256 constant AGGREGATION_RING_SIZE = 128;
+uint128 constant AGGREGATION_RING_SIZE = 128;
